LayoutTests/imported/w3c/web-platform-tests/css/css-color/nested-color-mix-with-currentcolor-expected.txt


FAIL Nested color-mix function with inner currentColor should inherit unresolved assert_equals: expected "lch(0 0 none)" but got "lch(0 0 0)"
